Reorganise functions into separate files as appropriate.
