+++ lib/Makefile
@@ -3,6 +3,14 @@ include ../config.mk
 .PHONY : really clean install
 
 MOSQ_OBJS=mosquitto.o \
+		  handle_connack.o \
+		  handle_ping.o \
+		  handle_pubackcomp.o \
+		  handle_publish.o \
+		  handle_pubrec.o \
+		  handle_pubrel.o \
+		  handle_suback.o \
+		  handle_unsuback.o \
 		  helpers.o \
 		  logging_mosq.o \
 		  memory_mosq.o \
@@ -10,10 +18,12 @@ MOSQ_OBJS=mosquitto.o \
 		  net_mosq.o \
 		  packet_mosq.o \
 		  read_handle.o \
-		  read_handle_client.o \
-		  read_handle_shared.o \
+		  send_connect.o \
+		  send_disconnect.o \
 		  send_mosq.o \
-		  send_client_mosq.o \
+		  send_publish.o \
+		  send_subscribe.o \
+		  send_unsubscribe.o \
 		  socks_mosq.o \
 		  srv_mosq.o \
 		  thread_mosq.o \
@@ -49,7 +59,32 @@ libmosquitto.so.${SOVERSION} : ${MOSQ_OBJS}
 
 libmosquitto.a : ${MOSQ_OBJS}
 	${CROSS_COMPILE}$(AR) cr $@ $^
+
 mosquitto.o : mosquitto.c mosquitto.h
+	${CROSS_COMPILE}$(CC) $(LIB_CFLAGS) -c $< -o $@
+
+handle_connack.o : handle_connack.c read_handle.h
+	${CROSS_COMPILE}$(CC) $(LIB_CFLAGS) -c $< -o $@
+
+handle_publish.o : handle_publish.c read_handle.h
+	${CROSS_COMPILE}$(CC) $(LIB_CFLAGS) -c $< -o $@
+
+handle_ping.o : handle_ping.c read_handle.h
+	${CROSS_COMPILE}$(CC) $(LIB_CFLAGS) -c $< -o $@
+
+handle_pubackcomp.o : handle_pubackcomp.c read_handle.h
+	${CROSS_COMPILE}$(CC) $(LIB_CFLAGS) -c $< -o $@
+
+handle_pubrec.o : handle_pubrec.c read_handle.h
+	${CROSS_COMPILE}$(CC) $(LIB_CFLAGS) -c $< -o $@
+
+handle_pubrel.o : handle_pubrel.c read_handle.h
+	${CROSS_COMPILE}$(CC) $(LIB_CFLAGS) -c $< -o $@
+
+handle_suback.o : handle_suback.c read_handle.h
+	${CROSS_COMPILE}$(CC) $(LIB_CFLAGS) -c $< -o $@
+
+handle_unsuback.o : handle_unsuback.c read_handle.h
 	${CROSS_COMPILE}$(CC) $(LIB_CFLAGS) -c $< -o $@
 
 helpers.o : helpers.c
@@ -73,16 +108,22 @@ packet_mosq.o : packet_mosq.c packet_mosq.h
 read_handle.o : read_handle.c read_handle.h
 	${CROSS_COMPILE}$(CC) $(LIB_CFLAGS) -c $< -o $@
 
-read_handle_client.o : read_handle_client.c read_handle.h
+send_connect.o : send_connect.c send_mosq.h
 	${CROSS_COMPILE}$(CC) $(LIB_CFLAGS) -c $< -o $@
 
-read_handle_shared.o : read_handle_shared.c read_handle.h
+send_disconnect.o : send_disconnect.c send_mosq.h
 	${CROSS_COMPILE}$(CC) $(LIB_CFLAGS) -c $< -o $@
 
 send_mosq.o : send_mosq.c send_mosq.h
 	${CROSS_COMPILE}$(CC) $(LIB_CFLAGS) -c $< -o $@
 
-send_client_mosq.o : send_client_mosq.c send_mosq.h
+send_publish.o : send_publish.c send_mosq.h
+	${CROSS_COMPILE}$(CC) $(LIB_CFLAGS) -c $< -o $@
+
+send_subscribe.o : send_subscribe.c send_mosq.h
+	${CROSS_COMPILE}$(CC) $(LIB_CFLAGS) -c $< -o $@
+
+send_unsubscribe.o : send_unsubscribe.c send_mosq.h
 	${CROSS_COMPILE}$(CC) $(LIB_CFLAGS) -c $< -o $@
 
 socks_mosq.o : socks_mosq.c
