+++ lib/handle_pubrec.c
@@ -4,12 +4,12 @@ Copyright (c) 2009-2015 Roger Light <roger@atchoo.org>
 All rights reserved. This program and the accompanying materials
 are made available under the terms of the Eclipse Public License v1.0
 and Eclipse Distribution License v1.0 which accompany this distribution.
- 
+
 The Eclipse Public License is available at
    http://www.eclipse.org/legal/epl-v10.html
 and the Eclipse Distribution License is available at
   http://www.eclipse.org/org/documents/edl-v10.php.
- 
+
 Contributors:
    Roger Light - initial implementation and documentation.
 */
@@ -27,38 +27,33 @@ Contributors:
 #include "packet_mosq.h"
 #include "read_handle.h"
 #include "send_mosq.h"
-#include "time_mosq.h"
 #include "util_mosq.h"
+#ifdef WITH_BROKER
+#include "mosquitto_broker.h"
+#endif
 
-int handle__packet(struct mosquitto *mosq)
+
+int handle__pubrec(struct mosquitto *mosq)
 {
-	assert(mosq);
+	uint16_t mid;
+	int rc;
 
-	switch((mosq->in_packet.command)&0xF0){
-		case PINGREQ:
-			return handle__pingreq(mosq);
-		case PINGRESP:
-			return handle__pingresp(mosq);
-		case PUBACK:
-			return handle__pubackcomp(mosq, "PUBACK");
-		case PUBCOMP:
-			return handle__pubackcomp(mosq, "PUBCOMP");
-		case PUBLISH:
-			return handle__publish(mosq);
-		case PUBREC:
-			return handle__pubrec(mosq);
-		case PUBREL:
-			return handle__pubrel(NULL, mosq);
-		case CONNACK:
-			return handle__connack(mosq);
-		case SUBACK:
-			return handle__suback(mosq);
-		case UNSUBACK:
-			return handle__unsuback(mosq);
-		default:
-			/* If we don't recognise the command, return an error straight away. */
-			log__printf(mosq, MOSQ_LOG_ERR, "Error: Unrecognised command %d\n", (mosq->in_packet.command)&0xF0);
-			return MOSQ_ERR_PROTOCOL;
-	}
+	assert(mosq);
+	rc = packet__read_uint16(&mosq->in_packet, &mid);
+	if(rc) return rc;
+#ifdef WITH_BROKER
+	log__printf(NULL, MOSQ_LOG_DEBUG, "Received PUBREC from %s (Mid: %d)", mosq->id, mid);
+
+	rc = db__message_update(mosq, mid, mosq_md_out, mosq_ms_wait_for_pubcomp);
+#else
+	log__printf(mosq, MOSQ_LOG_DEBUG, "Client %s received PUBREC (Mid: %d)", mosq->id, mid);
+
+	rc = message__out_update(mosq, mid, mosq_ms_wait_for_pubcomp);
+#endif
+	if(rc) return rc;
+	rc = send__pubrel(mosq, mid);
+	if(rc) return rc;
+
+	return MOSQ_ERR_SUCCESS;
 }
