+++ src/Makefile
@@ -8,7 +8,50 @@ else
 all : mosquitto
 endif
 
-mosquitto : mosquitto.o bridge.o conf.o context.o database.o logging.o loop.o memory_mosq.o persist.o net.o net_mosq.o packet_mosq.o read_handle.o read_handle_client.o read_handle_server.o read_handle_shared.o security.o security_default.o send_client_mosq.o send_mosq.o send_server.o service.o subs.o sys_tree.o time_mosq.o tls_mosq.o util_mosq.o websockets.o will_mosq.o
+OBJS=	mosquitto.o \
+		bridge.o \
+		conf.o \
+		context.o \
+		database.o \
+		handle_connack.o \
+		handle_connect.o \
+		handle_ping.o \
+		handle_pubackcomp.o \
+		handle_publish.o \
+		handle_pubrec.o \
+		handle_pubrel.o \
+		handle_suback.o \
+		handle_subscribe.o \
+		handle_unsuback.o \
+		handle_unsubscribe.o \
+		logging.o \
+		loop.o \
+		memory_mosq.o \
+		persist.o \
+		net.o \
+		net_mosq.o \
+		packet_mosq.o \
+		read_handle.o \
+		security.o \
+		security_default.o \
+		send_connack.o \
+		send_connect.o \
+		send_disconnect.o \
+		send_mosq.o \
+		send_publish.o \
+		send_suback.o \
+		send_subscribe.o \
+		send_unsubscribe.o \
+		service.o \
+		subs.o \
+		sys_tree.o \
+		time_mosq.o \
+		tls_mosq.o \
+		util_mosq.o \
+		websockets.o \
+		will_mosq.o
+
+mosquitto : ${OBJS}
 	${CROSS_COMPILE}${CC} $^ -o $@ ${LDFLAGS} $(BROKER_LIBS)
 
 mosquitto.o : mosquitto.c mosquitto_broker.h
@@ -24,6 +67,39 @@ context.o : context.c mosquitto_broker.h
 	${CROSS_COMPILE}${CC} $(BROKER_CFLAGS) -c $< -o $@
 
 database.o : database.c mosquitto_broker.h
+	${CROSS_COMPILE}${CC} $(BROKER_CFLAGS) -c $< -o $@
+
+handle_connack.o : handle_connack.c mosquitto_broker.h
+	${CROSS_COMPILE}${CC} $(BROKER_CFLAGS) -c $< -o $@
+
+handle_connect.o : handle_connect.c mosquitto_broker.h
+	${CROSS_COMPILE}${CC} $(BROKER_CFLAGS) -c $< -o $@
+
+handle_ping.o : ../lib/handle_ping.c ../lib/read_handle.h
+	${CROSS_COMPILE}${CC} $(BROKER_CFLAGS) -c $< -o $@
+
+handle_pubackcomp.o : ../lib/handle_pubackcomp.c ../lib/read_handle.h
+	${CROSS_COMPILE}${CC} $(BROKER_CFLAGS) -c $< -o $@
+
+handle_publish.o : handle_publish.c mosquitto_broker.h
+	${CROSS_COMPILE}${CC} $(BROKER_CFLAGS) -c $< -o $@
+
+handle_pubrec.o : ../lib/handle_pubrec.c ../lib/read_handle.h
+	${CROSS_COMPILE}${CC} $(BROKER_CFLAGS) -c $< -o $@
+
+handle_pubrel.o : ../lib/handle_pubrel.c ../lib/read_handle.h
+	${CROSS_COMPILE}${CC} $(BROKER_CFLAGS) -c $< -o $@
+
+handle_suback.o : ../lib/handle_suback.c ../lib/read_handle.h
+	${CROSS_COMPILE}${CC} $(BROKER_CFLAGS) -c $< -o $@
+
+handle_subscribe.o : handle_subscribe.c mosquitto_broker.h
+	${CROSS_COMPILE}${CC} $(BROKER_CFLAGS) -c $< -o $@
+
+handle_unsuback.o : ../lib/handle_unsuback.c ../lib/read_handle.h
+	${CROSS_COMPILE}${CC} $(BROKER_CFLAGS) -c $< -o $@
+
+handle_unsubscribe.o : handle_unsubscribe.c mosquitto_broker.h
 	${CROSS_COMPILE}${CC} $(BROKER_CFLAGS) -c $< -o $@
 
 logging.o : logging.c mosquitto_broker.h
@@ -50,28 +126,34 @@ packet_mosq.o : ../lib/packet_mosq.c ../lib/packet_mosq.h
 read_handle.o : read_handle.c mosquitto_broker.h
 	${CROSS_COMPILE}${CC} $(BROKER_CFLAGS) -c $< -o $@
 
-read_handle_client.o : read_handle_client.c mosquitto_broker.h
-	${CROSS_COMPILE}${CC} $(BROKER_CFLAGS) -c $< -o $@
-
-read_handle_server.o : read_handle_server.c mosquitto_broker.h
-	${CROSS_COMPILE}${CC} $(BROKER_CFLAGS) -c $< -o $@
-
-read_handle_shared.o : ../lib/read_handle_shared.c ../lib/read_handle.h
-	${CROSS_COMPILE}${CC} $(BROKER_CFLAGS) -c $< -o $@
-
 security.o : security.c mosquitto_broker.h
 	${CROSS_COMPILE}${CC} $(BROKER_CFLAGS) -c $< -o $@
 
 security_default.o : security_default.c mosquitto_broker.h
 	${CROSS_COMPILE}${CC} $(BROKER_CFLAGS) -c $< -o $@
 
-send_client_mosq.o : ../lib/send_client_mosq.c ../lib/send_mosq.h
+send_connect.o : ../lib/send_connect.c ../lib/send_mosq.h
+	${CROSS_COMPILE}${CC} $(BROKER_CFLAGS) -c $< -o $@
+
+send_disconnect.o : ../lib/send_disconnect.c ../lib/send_mosq.h
+	${CROSS_COMPILE}${CC} $(BROKER_CFLAGS) -c $< -o $@
+
+send_connack.o : send_connack.c mosquitto_broker.h
 	${CROSS_COMPILE}${CC} $(BROKER_CFLAGS) -c $< -o $@
 
 send_mosq.o : ../lib/send_mosq.c ../lib/send_mosq.h
 	${CROSS_COMPILE}${CC} $(BROKER_CFLAGS) -c $< -o $@
 
-send_server.o : send_server.c mosquitto_broker.h
+send_publish.o : ../lib/send_publish.c ../lib/send_mosq.h
+	${CROSS_COMPILE}${CC} $(BROKER_CFLAGS) -c $< -o $@
+
+send_suback.o : send_suback.c mosquitto_broker.h
+	${CROSS_COMPILE}${CC} $(BROKER_CFLAGS) -c $< -o $@
+
+send_subscribe.o : ../lib/send_subscribe.c ../lib/send_mosq.h
+	${CROSS_COMPILE}${CC} $(BROKER_CFLAGS) -c $< -o $@
+
+send_unsubscribe.o : ../lib/send_unsubscribe.c ../lib/send_mosq.h
 	${CROSS_COMPILE}${CC} $(BROKER_CFLAGS) -c $< -o $@
 
 service.o : service.c mosquitto_broker.h
